made enemy class file
