+++ test_platformer.py
@@ -2,9 +2,11 @@
 Platformer Template
 """
 import arcade
+import math
 import launch
 import random
 from mario import Mario
+from enemy import Enemy
 import json
 from mystery_box import Mystery_Box
 from coin import Coin
@@ -17,6 +19,9 @@ SCREEN_TITLE = "Platformer"
 SCREEN_WIDTH = 640
 SCREEN_HEIGHT = 600
 DEFAULT_FONT_SIZE = 25
+
+# Constants used to track if the player is facing left or right
+RIGHT_FACING = 0
 
 # Constants used to scale our sprites from their original size
 CHARACTER_SCALING = 2.5
@@ -41,67 +46,6 @@ LAYER_NAME_COINS = "Coins"
 LAYER_NAME_BACKGROUND = "Background"
 LAYER_NAME_PLAYER = "Player"
 LAYER_NAME_ENEMIES = "Enemies"
-
-class Entity(arcade.Sprite):
-    def __init__(self, name_folder, name_file):
-        super().__init__()
-
-        # Default to facing right
-        self.facing_direction = RIGHT_FACING
-
-        # Used for image sequences
-        self.cur_texture = 0
-        self.scale = CHARACTER_SCALING
-        self.character_face_direction = RIGHT_FACING
-
-        main_path = f":resources:images/animated_characters/{name_folder}/{name_file}"
-
-        self.idle_texture_pair = load_texture_pair(f"{main_path}_idle.png")
-        self.jump_texture_pair = load_texture_pair(f"{main_path}_jump.png")
-        self.fall_texture_pair = load_texture_pair(f"{main_path}_fall.png")
-
-        # Load textures for walking
-        self.walk_textures = []
-        for i in range(8):
-            texture = load_texture_pair(f"{main_path}_walk{i}.png")
-            self.walk_textures.append(texture)
-
-        # Load textures for climbing
-        self.climbing_textures = []
-        texture = arcade.load_texture(f"{main_path}_climb0.png")
-        self.climbing_textures.append(texture)
-        texture = arcade.load_texture(f"{main_path}_climb1.png")
-        self.climbing_textures.append(texture)
-
-        # Set the initial texture
-        self.texture = self.idle_texture_pair[0]
-
-        # Hit box will be set based on the first image used. If you want to specify
-        # a different hit box, you can do it like the code below.
-        # set_hit_box = [[-22, -64], [22, -64], [22, 28], [-22, 28]]
-        self.hit_box = self.texture.hit_box_points
-
-
-class Enemy(Entity):
-    def __init__(self, name_folder, name_file):
-
-        # Setup parent class
-        super().__init__(name_folder, name_file)
-
-
-class GoombaEnemy(Enemy):
-    def __init__(self):
-
-        # Set up parent class
-        super().__init__("goomba", "goomba")
-
-
-#class ZombieEnemy(Enemy):
-    #def __init__(self):
-
-        # Set up parent class
-        #super().__init__("zombie", "zombie")
-
 
 
 class MyGame(arcade.Window):
